AFUNK3_5-17-2025:: Refactor datatypes folder.
