+++ app/src/main/java/com/example/arbenchapp/datatypes/postprocessing/ConversionMethod.java
@@ -1,4 +1,4 @@
-package com.example.arbenchapp.datatypes;
+package com.example.arbenchapp.datatypes.postprocessing;
 
 public enum ConversionMethod {
     DEFAULT,
