+++ app/src/main/java/com/example/arbenchapp/datatypes/preprocessing/ModelType.java
@@ -1,4 +1,4 @@
-package com.example.arbenchapp.datatypes;
+package com.example.arbenchapp.datatypes.preprocessing;
 
 public enum ModelType {
     ERROR,
